Refactor : gameController를 사용하여 핸들러 리팩토링
상세 :
- gameController의 함수 호출시 gameManager와 timer를 전달
