+++ server/service/game/eventHandlers/sendChattingMessageHandler.js
@@ -1,6 +1,18 @@
 const short = require('short-uuid');
 const { io } = require('../../io');
-const roomController = require('../roomController');
+const roomController = require('../controllers/roomController');
+const gameController = require('../controllers/gameController');
+
+/**
+ * viewer가 입력한 채팅이 정답이라면 true를 반환하는 함수
+ */
+const isCorrectAnswer = (gameManager, message, socketId) => {
+  return (
+    gameManager.getStatus() === 'playing' &&
+    gameManager.getQuiz() === message &&
+    !gameManager.isStreamer(socketId)
+  );
+};
 
 const sendChattingMessageHandler = (socket, { nickname, message }) => {
   /**
@@ -8,14 +20,10 @@ const sendChattingMessageHandler = (socket, { nickname, message }) => {
    * room.quiz === message
    *  emit('정답!')
    */
-  const { gameManager } = roomController.getRoomByRoomId(socket.roomId);
+  const { gameManager, timer } = roomController.getRoomByRoomId(socket.roomId);
   const player = gameManager.getPlayerBySocketId(socket.id);
 
-  if (
-    gameManager.getStatus() === 'playing' &&
-    gameManager.getQuiz() === message &&
-    !gameManager.isStreamer(socket.id)
-  ) {
+  if (isCorrectAnswer(gameManager, message, socket.id)) {
     io.in(socket.roomId).emit('sendChattingMessage', {
       nickname: '안내',
       message: `${nickname}님이 정답을 맞췄습니다!`,
@@ -28,21 +36,17 @@ const sendChattingMessageHandler = (socket, { nickname, message }) => {
     player.setIsCorrectPlayer(true);
 
     if (gameManager.checkAllPlayersAreCorrect()) {
-      io.in(gameManager.getRoomId()).emit('endSet', {
-        scoreList: gameManager.getScoreList(),
-      });
-      gameManager.reset();
-      gameManager.clearPlayingTimer();
-      gameManager.resetAllPlayers();
+      gameController.endSet(gameManager, timer);
     }
-  } else {
-    io.in(socket.roomId).emit('sendChattingMessage', {
-      nickname,
-      message,
-      nicknameColor: player.getNicknameColor(),
-      id: short.generate(),
-    });
+    return;
   }
+
+  io.in(socket.roomId).emit('sendChattingMessage', {
+    nickname,
+    message,
+    nicknameColor: player.getNicknameColor(),
+    id: short.generate(),
+  });
 };
 
 module.exports = sendChattingMessageHandler;
